Fix and modernize todomvc.
This makes it so we can test it with the new package bot scripts. This moves
shared code to lib folder and correctly refers to shared files from tests. We
also switch to use core elements (we had a clone of them from eons ago).

Review URL: https://codereview.chromium.org//682373002